Crate defmt_or_log

Modules

macros

wrappers

Newtypes that implement styles of display that can not be expressed with a unified portable expression in both defmt and Rust string formatting.

Macros

assert

assert_eq

assert_ne

debug

debug_assert

debug_assert_eq

debug_assert_ne

error

expect

info

intern

panic

todo

trace

unimplemented

unreachable

unwrap

warn

Structs

Debug2Format

Copy of defmt::Debug2Format, except it behaves transparent if log is enabled

Display2Format

Copy of defmt::Display2Format, except it behaves transparent if log is enabled

Traits

FormatOrDebug

A trait that describes a type which can be logged with defmt or log, depending on the current configuration.

Attribute Macros

derive_format_or_debug

maybe_derive_debug

maybe_derive_format